Move `processed_ds_urls` to the containing conftest.py
