+++ datalad_registry/tests/test_tasks/conftest.py
@@ -1,6 +1,7 @@
 import pytest
 
 from datalad_registry.models import URL, db
+from datalad_registry.tasks import process_dataset_url
 
 from . import TEST_MIN_REPO_URL
 
@@ -31,3 +32,23 @@ def populate_db_with_unprocessed_dataset_urls(
 
         db.session.delete(non_dataset_url)
         db.session.commit()  # 1 is no longer a valid dataset URL id
+
+
+@pytest.fixture
+def processed_ds_urls(flask_app, two_files_ds_annex) -> list[int]:
+    """
+    Add valid dataset URLs to the database, process them, and return their IDs,
+    the primary keys
+    """
+
+    urls = [URL(url=TEST_MIN_REPO_URL), URL(url=two_files_ds_annex.path)]
+
+    with flask_app.app_context():
+        for url in urls:
+            db.session.add(url)
+        db.session.commit()
+
+        for url in urls:
+            process_dataset_url(url.id)
+
+        return [url.id for url in urls]
