+++ datalad_registry/tests/test_tasks/test_extract_ds_meta.py
@@ -1,31 +1,9 @@
 import pytest
 
 from datalad_registry.models import URL, URLMetadata, db
-from datalad_registry.tasks import extract_ds_meta, process_dataset_url
-
-from . import TEST_MIN_REPO_URL
+from datalad_registry.tasks import extract_ds_meta
 
 _BASIC_EXTRACTOR = "metalad_core"
-
-
-@pytest.fixture
-def processed_ds_urls(flask_app, two_files_ds_annex) -> list[int]:
-    """
-    Add valid dataset URLs to the database, process them, and return their IDs,
-    the primary keys
-    """
-
-    urls = [URL(url=TEST_MIN_REPO_URL), URL(url=two_files_ds_annex.path)]
-
-    with flask_app.app_context():
-        for url in urls:
-            db.session.add(url)
-        db.session.commit()
-
-        for url in urls:
-            process_dataset_url(url.id)
-
-        return [url.id for url in urls]
 
 
 class TestExtractDsMeta:
